requirements.txt
python_pptx==0.6.18
pygame==1.9.6
Pillow==6.2.1
PyQt5==5.14.0
